requirements.txt
rich
python-nmap

